TextRenderer: Fix shadow and styling offset in unicode mode

+++ src/main/java/fi/dy/masa/malilib/render/RenderUtils.java
@@ -154,7 +154,7 @@ public class RenderUtils
         buffer.pos(x + width, y         , zLevel).color(r, g, b, a).endVertex();
     }
 
-    public static void renderRectangleBatched(int x, int y, float z, int width, int height, Color4f color, BufferBuilder buffer)
+    public static void renderRectangleBatched(float x, float y, float z, int width, int height, Color4f color, BufferBuilder buffer)
     {
         buffer.pos(x        , y         , z).color(color.r, color.g, color.b, color.a).endVertex();
         buffer.pos(x        , y + height, z).color(color.r, color.g, color.b, color.a).endVertex();
